Rename SignalWire class to RelayClient for consistency with client exports
