+++ packages/node/index.ts
@@ -1,7 +1,7 @@
 require('dotenv').config()
 
 import RestClient from './src/rest'
-import RelayClient from './src/SignalWire'
+import RelayClient from './src/relay'
 import { setWebSocket } from '../common/src/services/Connection'
 
 import { Client } from 'faye-websocket'
