+++ packages/node/src/relay/index.ts
@@ -1,9 +1,8 @@
-import BaseSession from '../../common/src/BaseSession'
-import { deRegisterAll } from '../../common/src/services/Handler'
-import Calling from '../../common/src/relay/calling/Calling'
-import Connection from '../../common/src/services/Connection'
+import BaseSession from '../../../common/src/BaseSession'
+import Calling from '../../../common/src/relay/calling/Calling'
+import Connection from '../../../common/src/services/Connection'
 
-export default class SignalWire extends BaseSession {
+export default class extends BaseSession {
   private _callingInstance: Calling = null
 
   async connect(): Promise<void> {
